requirements.txt
aiohttp
aiohttp_jinja2
google-cloud-pubsub
jinja2
websockets
protobuf==6.31.0 # Match the gencode version from the error
redis
